ENG-0000 - Implicit Residency Awareness
Slightly tweaks the behavior of the base API client's `prepare()`
function to use residency-aware lookups for specific residency zones,
even if they are not in the default list of residency-aware services.
This allows the original default behavior to remain, but lets individual
client methods choose their type of residency behavior on a case-by-case
basis.