ENG-0000 - Add Runtime Configuration for Link Truncation
Defaults to off.  It is is enable, the routing layer will reduce any
links it generates to itself to only the anchor fragment and query
string.  Hypothetically, this should universally be the case -- and if
it auditions well in the embedded fortra-platform version of magma it
should soon become so.